+++ src/components/MainLayout/AppContent/AppContent.jsx
@@ -1,6 +1,7 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
 
 import Accounts from '@/views/Accounts';
+import Categories from '@/views/Categories';
 import Dashboard from '@/views/Dashboard';
 import Home from '@/views/Home';
 import Ledger from '@/views/Ledger';
@@ -23,6 +24,10 @@ export default function AppContent() {
       <Route
         path='/accounts/:accountId'
         element={<Ledger />}
+      />
+      <Route
+        path='/categories'
+        element={<Categories />}
       />
       <Route
         path='*'
