+++ src/views/Categories/Categories.jsx
@@ -1,0 +1,7 @@
+export default function Categories() {
+  return (
+    <div>
+      <h1>Categories</h1>
+    </div>
+  );
+}
